# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alina group test project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Authorization",
                "operationId": "login",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "email@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "data": {
                                            "id": 1,
                                            "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0L2FwaS92MS9sb2dpbiIsImlhdCI6MTcxMzYzNjQ0NywiZXhwIjoxNzEzNjQwMDQ3LCJuYmYiOjE3MTM2MzY0NDcsImp0aSI6Inh1VnhEalVUb1Y3c043R2kiLCJzdWIiOiIxIiwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.WzOOaWiaT-mDTzth9TiXvOpudUaEArqMy1fadbpkk6A",
                                            "name": "name",
                                            "surname": "surname",
                                            "email": "email@test.com",
                                            "updated_at": "2024-04-20T18:07:27.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Incorrect login or password"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "User not found"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registration",
                "operationId": "register",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Name"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Surname"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "email1@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "data": {
                                            "id": 1,
                                            "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0L2FwaS92MS9sb2dpbiIsImlhdCI6MTcxMzYzNjQ0NywiZXhwIjoxNzEzNjQwMDQ3LCJuYmYiOjE3MTM2MzY0NDcsImp0aSI6Inh1VnhEalVUb1Y3c043R2kiLCJzdWIiOiIxIiwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.WzOOaWiaT-mDTzth9TiXvOpudUaEArqMy1fadbpkk6A",
                                            "name": "name",
                                            "surname": "surname",
                                            "email": "email@test.com",
                                            "updated_at": "2024-04-20T18:07:27.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Incorrect login or password"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "User not found"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}